// src/app/generated/commander/model/commandRequest.ts
/**
 * plexhelp-commander API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CommandRequest { 
    name?: string;
    params?: Array<string> | null;
}