// src/app/generated/commander/model/status.ts
/**
 * plexhelp-commander API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type Status = 'SUCCESS' | 'NOT_FOUND' | 'FAILED';

export const Status = {
    Success: 'SUCCESS' as Status,
    NotFound: 'NOT_FOUND' as Status,
    Failed: 'FAILED' as Status
};